requirements.txt
tritonclient[grpc]
fastapi[standard]
httpx
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyyaml
uvicorn
bcrypt
sqlalchemy
psycopg2-binary
alembic
PyJWT